requirements.txt
flask
flask-wtf
flask-testing
python-dotenv
flask-sqlalchemy
flask-restful
flask-marshmallow
marshmallow-sqlalchemy
python-decouple